@@ -3,8 +3,8 @@
 namespace Database\Factories;
 
 use App\Models\User;
-use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Str;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
 class UserFactory extends Factory
 {
@@ -24,10 +24,25 @@ class UserFactory extends Factory
     {
         return [
             'name' => $this->faker->name,
+            'username' => $this->faker->unique()->userName,
             'email' => $this->faker->unique()->safeEmail,
+            'phone' => $this->faker->unique()->phoneNumber,
             'email_verified_at' => now(),
             'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
             'remember_token' => Str::random(10),
+            'settings' => [],
+            'address' => [
+                'line1' => $this->faker->streetName,
+                'line2' => null,
+                'city' => $this->faker->city,
+                'state' => $this->faker->state,
+                'country' => $this->faker->country,
+                'postal_code' => $this->faker->postcode,
+            ],
+            'locked' => false,
+            'profile_photo_path' => null,
+            'two_factor_secret' => null,
+            'two_factor_recovery_codes' => null,
         ];
     }
 }
