@@ -1,16 +1,19 @@
 <?php
 
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\Api\RunLogsController;
-use App\Http\Controllers\Api\StudentController;
 
-/* Test Route */
-Route::middleware(['api', 'ext-app'])->get('/ping', fn () => 'ping');
+/*
+|--------------------------------------------------------------------------
+| API Routes
+|--------------------------------------------------------------------------
+|
+| Here is where you can register API routes for your application. These
+| routes are loaded by the RouteServiceProvider within a group which
+| is assigned the "api" middleware group. Enjoy building your API!
+|
+*/
 
-Route::middleware(['api', 'ext-app'])->get(
-    '/students', [StudentController::class, 'index']
-);
-
-Route::middleware(['api', 'ext-app'])->post(
-    '/log/run', [RunLogsController::class, 'store']
-);
+Route::middleware('auth:api')->get('/user', function (Request $request) {
+    return $request->user();
+});
