@@ -1,11 +1,34 @@
 const defaultTheme = require('tailwindcss/defaultTheme');
 
 module.exports = {
-    purge: ['./storage/framework/views/*.php', './resources/views/**/*.blade.php'],
+    purge: [
+        './storage/framework/views/*.php',
+        './resources/views/**/*.blade.php',
+        './resources/js/**/*.vue',
+    ],
 
-    variants: {
-        opacity: ['responsive', 'hover', 'focus', 'disabled'],
+    darkMode: false, // or 'media' or 'class'
+
+    theme: {
+        container: {
+            center: true,
+        },
+
+        extend: {
+            fontFamily: {
+                sans: ['Inter', ...defaultTheme.fontFamily.sans],
+            },
+        },
     },
 
-    plugins: [require('@tailwindcss/ui')],
+    variants: {
+        extend: {
+            opacity: ['disabled'],
+        },
+    },
+
+    plugins: [
+        require('@tailwindcss/forms'),
+        require('@tailwindcss/typography'),
+    ],
 };
